Move hash types and LeafVersion to taproot-primitives
The hash types and LeafVersion enums are required by functionality in
the crypto crate. In order to facilitate the further separation of
logic into the crypto crate, these types need to be split from the
bitcoin crate.
