+++ taproot-primitives/CHANGELOG.md
@@ -1,0 +1,3 @@
+# 0.1.0 - Initial release
+
+- All the hash types and also the `LeafVersion` enum.
